dapp-refactor: moved ITransactionResult && ISmartContractExecutionResult to a shared type folder
