+++ system-contract-dapp-playground/__tests__/hedera/helper/index.test.ts
@@ -23,10 +23,6 @@ import {
   DEFAULT_IHTS_KEY_VALUE,
   KEY_TYPE_MAP,
 } from '@/utils/contract-interactions/HTS/token-create-custom/constant';
-import {
-  IHederaTokenServiceKeyType,
-  IHederaTokenServiceKeyValueType,
-} from '@/types/contract-interactions/HTS';
 
 describe('constructIHederaTokenKey test suite', () => {
   // mock contractId & compressedPubKey
