+++ system-contract-dapp-playground/src/components/common/methods/handleSanitizeFormInputs.tsx
@@ -19,7 +19,6 @@
  */
 
 import { isAddress } from 'ethers';
-import { ICommonKeyObject } from '@/types/contract-interactions/HTS';
 import { isCompressedPublicKey } from '@/utils/contract-interactions/HTS/helpers';
 
 interface ParamsProps {
