+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-721/methods/approve/index.tsx
@@ -24,8 +24,8 @@ import { Contract } from 'ethers';
 import { isAddress } from 'ethers';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { ReactNode, useCallback, useEffect, useState } from 'react';
-import { ITransactionResult } from '@/types/contract-interactions/HTS';
 import { erc721TokenApprove } from '@/api/hedera/erc721-interactions';
+import { ITransactionResult } from '@/types/contract-interactions/shared';
 import MultiLineMethod from '@/components/common/components/MultiLineMethod';
 import { handleAPIErrors } from '@/components/common/methods/handleAPIErrors';
 import HederaCommonTextField from '@/components/common/components/HederaCommonTextField';
