+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-721/methods/mint/index.tsx
@@ -24,7 +24,7 @@ import { useEffect, useState } from 'react';
 import { Contract, isAddress } from 'ethers';
 import { erc721Mint } from '@/api/hedera/erc721-interactions';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
-import { ITransactionResult } from '@/types/contract-interactions/HTS';
+import { ITransactionResult } from '@/types/contract-interactions/shared';
 import MultiLineMethod from '@/components/common/components/MultiLineMethod';
 import { handleAPIErrors } from '@/components/common/methods/handleAPIErrors';
 import { mintParamFields } from '@/utils/contract-interactions/erc/erc721/constant';
