+++ system-contract-dapp-playground/src/components/contract-interaction/hts/shared/components/TokenSpecificInfoModal.tsx
@@ -22,7 +22,6 @@ import Link from 'next/link';
 import { Dispatch, SetStateAction } from 'react';
 import { convertCalmelCaseFunctionName } from '@/utils/common/helpers';
 import { prepareInfoValuesToShow } from '../../../../common/methods/prepareInfoValuesToShow';
-import { IHederaTokenServiceKeyType } from '@/types/contract-interactions/HTS';
 import {
   HEDERA_BRANDING_COLORS,
   HEDERA_CHAKRA_TABLE_VARIANTS,
