+++ system-contract-dapp-playground/src/components/contract-interaction/hts/shared/methods/signingKeys.tsx
@@ -19,7 +19,6 @@
  */
 
 import { Dispatch, SetStateAction } from 'react';
-import { IHederaTokenServiceKeyType, ICommonKeyObject } from '@/types/contract-interactions/HTS';
 
 /** @dev handle updating number of keys*/
 export const handleAddingOrRemovingKeys = (
