+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-management-contract/methods/manageTokenStatus/index.tsx
@@ -23,7 +23,7 @@ import { useEffect, useState } from 'react';
 import { useToast } from '@chakra-ui/react';
 import { Contract, isAddress } from 'ethers';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
-import { ITransactionResult } from '@/types/contract-interactions/HTS';
+import { ITransactionResult } from '@/types/contract-interactions/shared';
 import { TRANSACTION_PAGE_SIZE } from '../../../shared/states/commonStates';
 import { handleAPIErrors } from '../../../../../common/methods/handleAPIErrors';
 import { useToastSuccessful } from '../../../../../../hooks/useToastSuccessful';
