+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-transfer-contract/method/transferCrypto/helpers/prepareCryptoTransferValues.ts
@@ -18,11 +18,6 @@
  *
  */
 
-import {
-  IHederaTokenServiceNftTransfer,
-  IHederaTokenServiceAccountAmount,
-  IHederaTokenServiceTokenTransferList,
-} from '@/types/contract-interactions/HTS';
 import { CryptoTransferParam, TokenTransferParam } from './generateInitialValues';
 
 interface CryptoTransferPageProps {
