+++ system-contract-dapp-playground/src/hooks/useUpdateLocalStorage.tsx
@@ -19,7 +19,7 @@
  */
 
 import { useEffect } from 'react';
-import { ITransactionResult } from '@/types/contract-interactions/HTS';
+import { ITransactionResult } from '@/types/contract-interactions/shared';
 
 /** @dev listen to change event on transactionResults state => load to localStorage */
 export const useUpdateTransactionResultsToLocalStorage = (
