+++ system-contract-dapp-playground/src/types/contract-interactions/HTS/index.d.ts
@@ -18,55 +18,6 @@
  *
  */
 
-/** @dev the type for HTS transaction results */
-export type ITransactionResult = {
-  status: 'success' | 'fail';
-  txHash: string;
-  APICalled?: any;
-  tokenInfo?: any;
-  isToken?: boolean;
-  keyTypeCalled?: any;
-  recordIndex?: number;
-  tokenAddress?: string;
-  initialAmount?: string;
-  accountAddress?: string;
-  transactionType: string;
-  convertedAmount?: string;
-  receiverAddress?: string;
-  pseudoRandomSeed?: string;
-  tokenAddresses?: string[];
-  transactionTimeStamp: number;
-  sessionedContractAddress: string;
-  transactionResultStorageKey: string;
-};
-
-/** @dev an interface for the results returned back from interacting with Hedera System Smart Contracts */
-interface ISmartContractExecutionResult {
-  Frozen?: any;
-  IsToken?: any;
-  Approved?: any;
-  TokenType?: any;
-  KycGranted?: any;
-  result?: boolean;
-  AllowanceValue?: any;
-  ApprovedAddress?: any;
-  tokenAddress?: string;
-  transactionHash?: string;
-  TokenDefaultKycStatus?: any;
-  TokenDefaultFreezeStatus?: any;
-  TokenInfo?: IHederaTokenServiceTokenInfo;
-  TokenKey?: IHederaTokenServiceKeyValueType;
-  TokenExpiryInfo?: IHederaTokenServiceExpiry;
-  FungibleTokenInfo?: IHederaTokenServiceFungibleTokenInfo;
-  NonFungibleTokenInfo?: IHederaTokenServiceNonFungibleTokenInfo;
-  TokenCustomFees?: {
-    fixedFees: IHederaTokenServiceFixedFee[];
-    royaltyFees: IHederaTokenServiceRoyaltyFee[];
-    fractionalFees: IHederaTokenServiceFractionalFee[];
-  };
-  err?: any;
-}
-
 /**
  * @dev a type for the IHederaTokenService.TokenKey.keyType
  *
